LG2
git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/external/LearnGreenlandic2@7944 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/common.hpp
@@ -5,7 +5,7 @@
 #include <utility>
 #include <set>
 
-const size_t lg2_revision = 7942;
+const uint lg2_revision = 7942;
 
 typedef std::pair<size_t,QString> dirmap_e;
 typedef std::set<dirmap_e,std::greater<dirmap_e> > dirmap_t;
